+++ example/lib/main0.dart
@@ -8,19 +8,18 @@ import 'package:get_it/get_it.dart';
 import 'package:get_state/get_state.dart';
 
 ///
-/// 示例0:
+/// 示例0:  本例仅供了解 get_state原理, 并不推荐本例中的写法
 /// 不使用injectable, 不使用View, 不使用自定义Model
 ///
 /// 注:
 ///   请 自下而上 查看本例源码
-///   本例仅供了解 get_state原理, 并不推荐下面的写法
 ///
 /// 1. 编写ViewModel类
 ///   ViewModel负责业务逻辑和操作视图
 ///   这里的操作Model的方法相当于BLoC中的Event
 ///
 /// 2. 在main方法中注册ViewModel,
-///   使用 GetIt g = GetIt.instance; 获取服务定位器实例
+///   使用 GetIt g = GetIt.instance; 获取GetIt实例
 ///   添加 WidgetsFlutterBinding.ensureInitialized(); 以防止ViewModel注册失败
 ///   使用 registerSingleton<T>(<构造方法>) 以懒单例的方式使用ViewModel
 ///   get_it 还有更多的注册方式, 这里暂时只介绍这一种
@@ -29,7 +28,7 @@ GetIt g = GetIt.instance;
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  // 4.手动依赖注入, 确保View可以获取到ViewModel
+  // 4.手动注入依赖, 确保View可以获取到ViewModel
   g.registerSingleton<CounterVm>(CounterVm());
   runApp(MyApp());
 }
@@ -64,7 +63,7 @@ class CounterVm extends ViewModel<int> {
   int counter()=> m;
 
   // 1.3 操作Model方法,
-  // 调用 父类中的vmRefresh()方法更新model的值
+  // 调用 父类中的vmUpdate()方法更新model的值
   void incrementCounter() {
     vmUpdate(m + 1);
   }